fix: rp-adapter validation of VP proofs
The RP Adapter is now signing the VP enclosing the authZ credential with
the tenant's DID before presenting it to the Issuer Adapter. The Issuer
Adapter is now verifying the proof on this VP.

The aries framework's `verifiable.ParsePresentation()` was fixed to
verify the embedded proofs in presentations:
hyperledger/aries-framework-go#2049.

Upgrading to the newest version exposed lots of bugs and assumptions
around the validation of proofs in VPs and VCs which have been addressed
in this patch, either as full fixes or followup action items.

Signed-off-by: George Aristy <george.aristy@securekey.com>