feat: issuer adapter may use access tokens scoped to the credential
If the connect/wallet endpoint is called without a `state` query param,
but with a `cred` query param containing a credential scope, then it
will request an access token scoped to openid, offline_access,
and the given cred scope, and with that access token being supplied
to the issuer with every request, the issuer can use the token to
determine response values.

The issuer profile must be created with at least one scope within the
`credScopes` list, to indicate which credential scopes the adapter
will be allowed to request from the issuer, for this to work.

Signed-off-by: Filip Burlacu <filip.burlacu@securekey.com>